refactor: relocate and standardize LanguageSwitcher component
